% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{Deployment}
\alias{Deployment}
\title{SKIL Deployment}
\usage{
Deployment(skil = NULL, name = NULL, deployment_id = NULL)
}
\arguments{
\item{skil:}{`Skil` server instance. If `None`, SKIL will load from config.}

\item{name:}{string. Name for the deployment.}

\item{id:}{Unique id for the deployment. If `None`, a unique id will be generated.ging.}
}
\description{
Deployments operate independently of workspaces to ensure that there are
no accidental interruptions or mistakes in a production environment.
}
\examples{
\dontrun{
library(skilr)

# Creating a deployment
deployment <- Deployment()

}
}
