% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{Experiment}
\alias{Experiment}
\title{SKIL Experiment}
\usage{
Experiment(work_space = NULL, experiment_id = NULL,
  name = "experiment", description = "experiment", verbose = FALSE,
  skil_server = NULL, create = TRUE)
}
\arguments{
\item{work_space:}{`WorkSpace` instance. If NULL a workspace will be created.}

\item{experiment_id:}{integer. Unique id for workspace. If `None`, a unique id will be generated.}

\item{name:}{string. Name for the experiment.}

\item{description:}{string. Description for the experiment.}

\item{verbose:}{boolean. If FALSE, api response will be printed.}

\item{skil_server:}{Optional `Skil` instance, used when create is false.}

\item{create:}{boolean. If TRUE a new experiment will be created.}
}
\description{
Experiments in SKIL are useful for defining different model configurations, 
encapsulating training of models, and carrying out different data cleaning tasks.
}
\details{
Experiments have a one-to-one relationship with Notebooks and have their own 
storage mechanism for saving different model configurations when seeking a best 
candidate.
}
\examples{
\dontrun{
library(skilr)

# Creating a SKIL experiment
skil <- Skil(host='localhost', port=9008, user_id='admin', password='admin')
ws   <- Workspace(skil, name='test_workspace')
exp  <- Experiment(ws, name='test_experiment')

}
}
