% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R, R/experiments.R
\name{get_experiment_by_id}
\alias{get_experiment_by_id}
\title{Get deployment by ID}
\usage{
get_experiment_by_id(skil, experiment_id)

get_experiment_by_id(skil, experiment_id)
}
\arguments{
\item{skil:}{`Skil` server instance}

\item{experiment_id:}{string, experiment ID}
}
\description{
skil: `Skil` server instance
deployment_id: string, deployment ID
}
