% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_model_by_id}
\alias{get_model_by_id}
\title{Get model by ID}
\usage{
get_model_by_id(experiment, model_id)
}
\arguments{
\item{experiment:}{SKIL experiment}

\item{model_id:}{string, model ID}
}
\description{
Get model by ID
}
