% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_transform_by_id}
\alias{get_transform_by_id}
\title{Get transform by ID}
\usage{
get_transform_by_id(experiment, model_id)
}
\arguments{
\item{transform_id:}{string, Transform ID}

\item{transform_type:}{string, Transfrom type ("CSV", "array" or "image")}

\item{experiment:}{SKIL Experiment instance}
}
\description{
Get transform by ID
}
