% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{get_workspace_by_id}
\alias{get_workspace_by_id}
\title{Get workspace by ID}
\usage{
get_workspace_by_id(skil, workspace_id)
}
\arguments{
\item{skil:}{`Skil` server instance}

\item{workspace_id:}{string, workspace ID}
}
\description{
Get workspace by ID
}
