% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Skil}
\alias{Skil}
\title{SKIL server connection}
\usage{
Skil(workspace_server_id = NULL, host = "localhost", port = 9008,
  user_id = "admin", password = "admin", debug = FALSE)
}
\arguments{
\item{workspace_server_id:}{None by default, only specify if you want to connect to a non-default SKIL workspace server.}

\item{host:}{string, Host on which the SKIL server runs.}

\item{port:}{integer, Port on which the SKIL host runs.}

\item{user_id:}{user name for your SKIL server connection.}

\item{password:}{password of the provided SKIL user.}

\item{debug:}{boolean, set to false for more verbose logging.}
}
\description{
Central class for managing connections with the SKIL server.
}
\examples{
\dontrun{
library(skilr)

# Connecting to SKIL
skil <- Skil(host='localhost', port=9008, user_id='admin', password='admin')

}
}
