% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{skil_from_config}
\alias{skil_from_config}
\title{Get SKIL server connection from configuration}
\usage{
skil_from_config()
}
\description{
Get SKIL server connection from configuration
}
\examples{
\dontrun{
library(skilr)

# Connecting to SKIL, assumes a prior connection from this machine
skil <- skil_from_config()

}
}
