% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{skilr-package}
\alias{skilr}
\alias{skilr-package}
\title{R interface to SKIL}
\description{
Train, deploy and manage your R models with SKIL
}
\author{
\strong{Maintainer}: Max Pumperla \email{max.pumperla@googlemail.com} (https://orcid.org/0000-0002-7801-4184)

}
