% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{TransformImageService}
\alias{TransformImageService}
\title{SKIL TransformImageService}
\usage{
TransformImageService(skil, model, deployment, model_entity)
}
\arguments{
\item{skil:}{`Skil` server instance}

\item{model:}{`skil.Model` instance}

\item{deployment:}{`skil.Deployment` instance}

\item{model_entity:}{skil_client.ModelEntity, result of `deploy_model` SKIL API call of a `Model`}
}
\description{
A service for transforming image data
}
\examples{
\dontrun{
library(skilr)
TODO

}
}
