% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{WorkSpace}
\alias{WorkSpace}
\title{SKIL Workspace}
\usage{
WorkSpace(skil = NULL, name = NULL, labels = NULL, verbose = FALSE,
  create = TRUE)
}
\arguments{
\item{skil:}{Skil server instance}

\item{name:}{string. Name for the workspace.}

\item{labels:}{string. Labels associated with the workspace, useful for searching (comma seperated).}

\item{verbose:}{boolean. If True, api response will be printed.}

\item{create:}{boolean. Internal, do not use.}
}
\description{
Workspaces are a collection of features that enable different tasks such as
conducting experiments, training models, and test different dataset transforms.
}
\details{
Workspaces are distinct from Deployments by operating as a space for 
non-production work.
}
\examples{
\dontrun{
library(skilr)

# Creating a SKIL workspace
skil <- Skil(host='localhost', port=9008, user_id='admin', password='admin')
ws   <- Workspace(skil, name='test_workspace')

}
}
